creacion de torneo y combate (aun falta)

+++ src/main/java/com/ipn/mx/demowebsocket/basedatos/domain/repository/AlumnoRepository.java
@@ -3,5 +3,5 @@ package com.ipn.mx.demowebsocket.basedatos.domain.repository;
 import com.ipn.mx.demowebsocket.basedatos.domain.entity.Alumno;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlumnoRepository extends JpaRepository<Alumno, Integer> {
+public interface AlumnoRepository extends JpaRepository<Alumno, Long> {
 }
