+++ src/main/java/com/ipn/mx/demowebsocket/basedatos/infrastructure/ParticipacionController.java
@@ -22,7 +22,7 @@ public class ParticipacionController {
     @GetMapping("/{idCombate}/{idAlumno}")
     @ResponseStatus(HttpStatus.OK)
     public Participacion read(@PathVariable Integer idCombate, @PathVariable Integer idAlumno) {
-        return service.read(new ParticipacionId(idCombate, idAlumno));
+        return service.read(new ParticipacionId(idCombate, idAlumno.longValue()));
     }
 
     @PostMapping
@@ -33,7 +33,7 @@ public class ParticipacionController {
     @ResponseStatus(HttpStatus.CREATED)
     public Participacion update(@PathVariable Integer idCombate, @PathVariable Integer idAlumno,
                                 @RequestBody Participacion p) {
-        Participacion actual = service.read(new ParticipacionId(idCombate, idAlumno));
+        Participacion actual = service.read(new ParticipacionId(idCombate, idAlumno.longValue()));
         actual.setCombate(p.getCombate());
         actual.setAlumno(p.getAlumno());
         return service.save(actual);
@@ -42,6 +42,6 @@ public class ParticipacionController {
     @DeleteMapping("/{idCombate}/{idAlumno}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void delete(@PathVariable Integer idCombate, @PathVariable Integer idAlumno) {
-        service.delete(new ParticipacionId(idCombate, idAlumno));
+        service.delete(new ParticipacionId(idCombate,  idAlumno.longValue()));
     }
 }
